дз3 пересдача

+++ Task002/src/Classes/Animals.java
@@ -1,17 +1,19 @@
 package Classes;
 
+import Classes.Interface.Actions;
+
 public class Animals implements Actions {
     public int height;
-    public int weight;
+    public double weight;
     public String colorEyes;
 
-    public Animals(int height, int weight, String colorEyes) {
+    public Animals(int height, double weight, String colorEyes) {
         this.height = height;
         this.weight = weight;
         this.colorEyes = colorEyes;
     }
     public void showInfo(){
-        System.out.printf("Рост: %d\nВес: %d\nЦвет глаз: %s", height, weight, colorEyes);
+        System.out.printf("Рост: %d\nВес: %f\nЦвет глаз: %s", height, weight, colorEyes);
     }
     public void voice(){
         System.out.println("Издать звук");
