+++ Task002/src/Classes/WildAnimals/FeaturesWildAnimals.java
@@ -6,7 +6,7 @@ public class FeaturesWildAnimals extends Animals {
     String habitat; // Место обитания
     String dateOfLocation; // Дата находждения
 
-    public FeaturesWildAnimals(int height, int weight, String colorEyes, String habitat, String dateOfLocation) {
+    public FeaturesWildAnimals(int height, double weight, String colorEyes, String habitat, String dateOfLocation) {
         super(height, weight, colorEyes);
         this.habitat = habitat;
         this.dateOfLocation = dateOfLocation;
